Create_Alignment_DB: fix fasta file parsing error
The > was previously used as input separator but some fasta headers in
uniref100 fasta contain >.

Also commented-out the AA statistics code and skipped any name cleaning.
These two take up most of the time.  The purpose of the AA stats is not
clear.  Some name-cleaning may be put back in the future after
performace is improved.