Create_Alignment_DB: skip all name cleaning (for now)
Since the CleanNames subroutine is universal, meaning it will scrub all
different kinds of "names" and has grown to be quite slow and lacks
documentation we take it out of commission for the time being.  Some
determination needs to be done later about this, maybe have dedicated
and speedier subroutines as a replacement.