CADB: move UR100 fasta and ID mapping input steps to the front
these steps don't depend (to the best of my knowledge) on the other
steps (but ID mapping depends on having the fasta data).  As the script
transistions towards running on partial/partitioned sets of input files,
the other steps may become more dependend on which UR100s or UPIDs were
loaded beforehand.